agents/requirements.txt
modal>=0.60.0
fastapi>=0.109.0
uvicorn>=0.27.0
anthropic>=0.40.0
firebase-admin>=6.4.0
qdrant-client>=1.7.0
python-telegram-bot>=21.0
PyGithub>=2.1.0
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
structlog>=24.1.0
exa-py>=1.0.0
tavily-python>=0.3.0
numpy>=1.24.0
PyYAML>=6.0.0

# Security & Performance (Phase 4)
slowapi>=0.1.9
aiofiles>=23.2.0

# Document Processing (Phase 3)
python-docx>=1.1.0
docxtpl>=0.16.0
docx2python>=3.0.0
openpyxl>=3.1.0
formulas>=1.2.0
python-pptx>=0.6.23
pypdf>=4.0.0
pdfplumber>=0.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

